Support automated build of installer and drivers
This commit adds "build.py" that supports the automated build of
the Windows installer and PV drivers.  It generates an output
directory "output/" containing the output files.  A README.md
is added to describe the dependencies and requirements of the
automated build.

build.py also supports "fetch" which retrieves all of the git
repositories necessary for the build.