generate C# branding with BeforeBuild build target
In C# SDK projects, the PreBuild target is not guaranteed to run every
build. It is also evaluated too late into the build, meaning that a
newly appearing branding file won't be accounted in time for the
compilation.

Use a separate target with BeforeTargets="BeforeBuild", and explicitly
include our branding file there.

Signed-off-by: Tu Dinh <ngoc-tu.dinh@vates.tech>